+++ src/js/source.js
@@ -2,23 +2,25 @@
 // Plyr source update
 // ==========================================================================
 
+import { providers, types } from './config/types';
 import html5 from './html5';
 import media from './media';
 import support from './support';
-import { providers, types } from './types';
 import ui from './ui';
-import utils from './utils';
+import { createElement, insertElement, removeElement } from './utils/elements';
+import is from './utils/is';
+import { getDeep } from './utils/objects';
 
 const source = {
     // Add elements to HTML5 media (source, tracks, etc)
     insertElements(type, attributes) {
-        if (utils.is.string(attributes)) {
-            utils.insertElement(type, this.media, {
+        if (is.string(attributes)) {
+            insertElement(type, this.media, {
                 src: attributes,
             });
-        } else if (utils.is.array(attributes)) {
+        } else if (is.array(attributes)) {
             attributes.forEach(attribute => {
-                utils.insertElement(type, this.media, attribute);
+                insertElement(type, this.media, attribute);
             });
         }
     },
@@ -26,7 +28,7 @@ const source = {
     // Update source
     // Sources are not checked for support so be careful
     change(input) {
-        if (!utils.is.object(input) || !('sources' in input) || !input.sources.length) {
+        if (!getDeep(input, 'sources.length')) {
             this.debug.warn('Invalid source format');
             return;
         }
@@ -42,10 +44,11 @@ const source = {
                 this.options.quality = [];
 
                 // Remove elements
-                utils.removeElement(this.media);
+                removeElement(this.media);
+                this.media = null;
 
                 // Reset class name
-                if (utils.is.element(this.elements.container)) {
+                if (is.element(this.elements.container)) {
                     this.elements.container.removeAttribute('class');
                 }
 
@@ -54,46 +57,25 @@ const source = {
                 const prevProvider = this.provider;
 
                 // Set the type and provider
-                this.type = input.type;
-                this.provider = !utils.is.empty(input.sources[0].provider) ? input.sources[0].provider : providers.html5;
+                const { sources, type } = input;
+                const [{ provider = providers.html5, src }] = sources;
+                const tagName = provider === 'html5' ? type : 'div';
+                const attributes = provider === 'html5' ? {} : { src };
 
-                // Check for support
-                this.supported = support.check(this.type, this.provider, this.config.playsinline);
-
-                // if provider and type is html5:video then to keep this.media to reuse
-                // because create new video tag will comsume a lot of memory in iOS hardware.
-                // iOS can only create about 16 video tag
-                if (prevType !== this.type || prevProvider !== this.provider || this.type !== types.video || this.provider !== providers.html5) {
-                    delete this.media;
-                    this.media = null;
-
-                    // Create new markup
-                    switch (`${this.provider}:${this.type}`) {
-                        case 'html5:video':
-                            this.media = utils.createElement('video');
-                            break;
-
-                        case 'html5:audio':
-                            this.media = utils.createElement('audio');
-                            break;
-
-                        case 'youtube:video':
-                        case 'vimeo:video':
-                            this.media = utils.createElement('div', {
-                                src: input.sources[0].src,
-                            });
-                            break;
-
-                        default:
-                            break;
-                    }
-                }
+                Object.assign(this, {
+                    provider,
+                    type,
+                    // Check for support
+                    supported: support.check(type, provider, this.config.playsinline),
+                    // Create new element
+                    media: createElement(tagName, attributes),
+                });
 
                 // Inject the new element
                 this.elements.container.appendChild(this.media);
 
                 // Autoplay the new source?
-                if (utils.is.boolean(input.autoplay)) {
+                if (is.boolean(input.autoplay)) {
                     this.config.autoplay = input.autoplay;
                 }
 
@@ -105,7 +87,7 @@ const source = {
                     if (this.config.autoplay) {
                         this.media.setAttribute('autoplay', '');
                     }
-                    if (!utils.is.empty(input.poster)) {
+                    if (!is.empty(input.poster)) {
                         this.poster = input.poster;
                     }
                     if (this.config.loop.active) {
@@ -124,7 +106,7 @@ const source = {
 
                 // Set new sources for html5
                 if (this.isHTML5) {
-                    source.insertElements.call(this, 'source', input.sources);
+                    source.insertElements.call(this, 'source', sources);
                 }
 
                 // Set video title